forum/src/main/java/com/example/forum/controller/AuthController.java
package com.example.forum.controller;

import com.example.forum.dto.CredentialsDTO;
import com.example.forum.security.JWTUtil;
import com.example.forum.service.UserDetailsServiceImpl;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/auth")
public class AuthController {

    @Autowired
    private AuthenticationManager authenticationManager;

    @Autowired
    private JWTUtil jwtUtil;

    @Autowired
    private UserDetailsServiceImpl userDetailsService;

    @PostMapping("/login")
    public ResponseEntity<?> login(@RequestBody CredentialsDTO credentials) {
        try {
            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(
                    credentials.getEmail(), credentials.getPassword()));
            UserDetails userDetails = userDetailsService.loadUserByUsername(credentials.getEmail());
            String token = jwtUtil.generateToken(userDetails.getUsername());
            return ResponseEntity.ok().header("Authorization", "Bearer " + token).build();
        } catch (AuthenticationException e) {
            return ResponseEntity.status(403).build(); // Unauthorized
        }
    }
}
